requirements.txt
# API
setuptools
wheel
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
gunicorn==21.2.0
streamlit==1.28.0

# YOLO
ultralytics==8.0.196

# OCR
easyocr==1.7.0

# Image processing
opencv-python-headless==4.8.1.78
Pillow==9.5.0
numpy==1.24.3
pandas==1.5.3

# Text matching
rapidfuzz==3.5.2

# Google Generative AI
google-generativeai==0.3.2
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0
